One should strive for perfection.
