+++ .github/workflows/development.yml
@@ -26,7 +26,7 @@ jobs:
         CI: true
 
 
-  lint:
+  lint vue:
 
     runs-on: ubuntu-latest
