+++ src/Views/CV/CV.vue
@@ -29,10 +29,12 @@
 
 <script>
 import cvSection from './cvSection/cvSection';
+import { skillSimplifierMixin } from '../../mixins/skillSimplifierMixin';
 
 import strings from '../../i18n.json';
 
 export default {
+	mixins: [ skillSimplifierMixin ],
 	components: { cvSection },
 	data: function() {
 		return {
@@ -40,23 +42,6 @@ export default {
 		};
 	},
 	computed: {
-		// This function destructures the skill objects to ignore the skill level
-		updatedSkills: function() {
-			let smallSkills = [];
-
-			for (let i = 0; i < strings.en.portfolio.skills.length; i++) {
-				let brandNewSkill = {};
-				brandNewSkill.name = strings.en.portfolio.skills[i].name;
-				brandNewSkill.description = [];
-
-				for (let j = 0; j < strings.en.portfolio.skills[i].description.length; j++) {
-
-					brandNewSkill.description.push(strings.en.portfolio.skills[i].description[j].name);
-				}
-				smallSkills.push(brandNewSkill);
-			}
-			return smallSkills;
-		},
 		restrictedAmountOfProject: function(){
 			const max3Project = strings.en.portfolio.projects.splice(0, 3) ;
