+++ src/Views/Portfolio/Skills/Skills.vue
@@ -19,10 +19,12 @@
 
 <script>
 import SkillCartridge from '../../../components/SkillCartridge/SkillCartridge';
+import { skillSimplifierMixin } from '../../../mixins/skillSimplifierMixin';
 
 import strings from '../../../i18n.json';
 
 export default {
+	mixins: [ skillSimplifierMixin ],
 	components: { SkillCartridge },
 	data: function() {
 		return {
@@ -30,23 +32,6 @@ export default {
 		};
 	},
 	computed: {
-		// This function destructures the skill objects to ignore the skill level
-		updatedSkills: function() {
-			let smallSkills = [];
-
-			for (let i = 0; i < strings.en.portfolio.skills.length; i++) {
-				let brandNewSkill = {};
-				brandNewSkill.name = strings.en.portfolio.skills[i].name;
-				brandNewSkill.description = [];
-
-				for (let j = 0; j < strings.en.portfolio.skills[i].description.length; j++) {
-
-					brandNewSkill.description.push(strings.en.portfolio.skills[i].description[j].name);
-				}
-				smallSkills.push(brandNewSkill);
-			}
-			return smallSkills;
-		},
 		restrictedAmountOfProject: function(){
 			const max3Project = strings.en.portfolio.projects.splice(0, 3) ;
